Refactored duplicate markdown resolution logic into a shared MarkdownUtils object.
Removed duplicate implementations of normalizeMarkdownImageSource, resolveMarkdownSourceUrl, extractImageSourcesFromText, and resolveOfflineMarkdownImages from CourseWizardActivity and DashboardCoursePageFragment.
