+++ test_script.kt
@@ -1,0 +1,5 @@
+import java.io.File
+
+fun main() {
+    println("Test script")
+}
